% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.r
\name{cookie}
\alias{cookie}
\title{Assemble an HTTP cookie.}
\usage{
cookie(
  cookie_value,
  max_age = NULL,
  domain = NULL,
  path = NULL,
  same_site = "Lax",
  secure = FALSE,
  http_only = FALSE,
  partitioned = FALSE
)
}
\arguments{
\item{cookie_value}{The value (string) to assign to the cookie.
Cannot contain white space, double quotes, commas,
semicolons, or slashes (\verb{" , ; \\\\}).}

\item{max_age}{The number of seconds until expiration.
Omit to create a session cookie,}

\item{domain}{Send with requests to this host.}

\item{path}{Send with requests to this path.}

\item{same_site}{\code{'Strict'}, \code{'Lax'}, or \code{'None'}.
\code{secure} required for \code{'None'}.}

\item{secure}{Only send over HTTPS.}

\item{http_only}{Disallow javascript access.}

\item{partitioned}{Use partitioned storage. \code{secure} required.}
}
\value{
A string.
}
\description{
See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie for
a more in-depth description of each parameter's purpose.
}
\examples{

    library(webqueue)
    
    cookie('xyz', max_age = 3600, http_only = TRUE)
    
    response(cookies = c(token = cookie('randomstring123')))
    
}
