% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.r
\name{js_obj}
\alias{js_obj}
\title{Ensure a list becomes a JSON object.}
\usage{
js_obj(x = list())
}
\arguments{
\item{x}{A list, or list-like object.}
}
\value{
A list with the names attribute set.
}
\description{
This function returns a list that \code{jsonlite::toJSON()} will always encode as
\code{{}}.
}
\examples{

    library(webqueue)
    
    updates <- list()
    
    response(list(updates = updates))
    
    response(list(updates = js_obj(updates)))
    
}
