% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{randomPort}
\alias{staticPath}
\alias{staticPathOptions}
\alias{fromJSON}
\alias{toJSON}
\alias{unbox}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{httpuv}{\code{\link[httpuv]{randomPort}}, \code{\link[httpuv]{staticPath}}, \code{\link[httpuv]{staticPathOptions}}}

  \item{jsonlite}{\code{\link[jsonlite]{fromJSON}}, \code{\link[jsonlite:fromJSON]{toJSON}}, \code{\link[jsonlite]{unbox}}}
}}

